backend/requirements.txt
# Core Framework
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6  # WebSocket support for real-time updates
python-socketio==5.10.0  # SocketIO client/server

# Authentication & Security
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0  # Note: Do NOT install 'jwt' package - it conflicts with PyJWT
bcrypt==4.1.2
Flask-Limiter==3.5.0  # Rate limiting
itsdangerous==2.1.2

# QR Code Generation
qrcode==7.4.2
Pillow==10.2.0  # For image processing in QR

# Utility Libraries
python-dotenv==1.0.1  # For .env file support
requests==2.31.0

# Database Support
psycopg2-binary==2.9.9  # For PostgreSQL (optional if using SQLite or MySQL)
pymysql==1.1.0          # For MySQL (optional)

# Logging & Monitoring
loguru==0.7.2

# Optional (for analytics, email, notifications)
# Note: smtplib is a built-in Python module, no need to install
pandas==2.2.0

# Testing
pytest==8.1.1
pytest-flask==1.3.0

# Compatibility & Performance
gunicorn==21.2.0  # For production server
